+++ pygeos/tests/test_measurement.py
@@ -7,6 +7,7 @@ import pygeos
 from .common import (
     empty,
     geometry_collection,
+    ignore_invalid,
     line_string,
     linear_ring,
     multi_line_string,
@@ -143,7 +144,10 @@ def test_hausdorff_distance():
     # example from GEOS docs
     a = pygeos.linestrings([[0, 0], [100, 0], [10, 100], [10, 100]])
     b = pygeos.linestrings([[0, 100], [0, 10], [80, 10]])
-    actual = pygeos.hausdorff_distance(a, b)
+    with ignore_invalid():
+        # Hausdorff distance emits "invalid value encountered"
+        # (see https://github.com/libgeos/geos/issues/515)
+        actual = pygeos.hausdorff_distance(a, b)
     assert actual == pytest.approx(22.360679775, abs=1e-7)
 
 
@@ -151,7 +155,10 @@ def test_hausdorff_distance_densify():
     # example from GEOS docs
     a = pygeos.linestrings([[0, 0], [100, 0], [10, 100], [10, 100]])
     b = pygeos.linestrings([[0, 100], [0, 10], [80, 10]])
-    actual = pygeos.hausdorff_distance(a, b, densify=0.001)
+    with ignore_invalid():
+        # Hausdorff distance emits "invalid value encountered"
+        # (see https://github.com/libgeos/geos/issues/515)
+        actual = pygeos.hausdorff_distance(a, b, densify=0.001)
     assert actual == pytest.approx(47.8, abs=0.1)